+++ lib/src/utils.dart
@@ -1,86 +1,61 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'layout_builder/layout_builder.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 
-/// Default ad layout
-AdLayoutBuilder get adBannerLayoutBuilder => (ratingBar, media, icon, headline,
-    advertiser, body, price, store, attribuition, button) {
-  return AdLinearLayout(
-    margin: EdgeInsets.all(10),
-    borderRadius: AdBorderRadius.all(10),
-    // The first linear layout width needs to be extended to the
-    // parents height, otherwise the children won't fit good
-    width: MATCH_PARENT,
-    children: [
-      AdLinearLayout(
-        children: [
-          icon,
-          AdLinearLayout(
-            children: [
-              headline,
-              AdLinearLayout(
-                children: [attribuition, advertiser],
-                orientation: HORIZONTAL,
-                width: WRAP_CONTENT,
-              ),
-            ],
-          ),
-        ],
-        width: MATCH_PARENT,
-        orientation: HORIZONTAL,
-        margin: EdgeInsets.all(6),
-      ),
-      button,
-    ],
-    backgroundColor: Colors.blue,
+void assertPlatformIsSupported() {
+  // Google Native ads are only supported in Android and iOS
+  assert(
+    Platform.isAndroid || Platform.isIOS,
+    'The current platform does not support native ads. The platforms that support it are Android and iOS',
   );
-};
 
-extension colorExtension on Color {
-  String toHex([bool hashtag = true]) {
-    if (this == Colors.transparent) return '#00ff0000';
-    String hex = '';
-    if (hashtag) hex = '#$hex';
-    hex = '$hex${value.toRadixString(16)}';
-    return hex;
-  }
+  assert(Platform.isAndroid, 'Android is the only supported platform for now');
 }
 
-class AdBorderRadius {
-  final double topRight, topLeft, bottomRight, bottomLeft;
+typedef AdBuilder = Widget Function(BuildContext context, Widget child);
 
-  AdBorderRadius({
-    this.topLeft,
-    this.topRight,
-    this.bottomLeft,
-    this.bottomRight,
-  });
-
-  static AdBorderRadius all(double value) => AdBorderRadius(
-        topLeft: value,
-        topRight: value,
-        bottomLeft: value,
-        bottomRight: value,
-      );
-
-  static AdBorderRadius vertical({
-    double top,
-    double bottom,
-  }) =>
-      AdBorderRadius(
-        topLeft: top,
-        topRight: top,
-        bottomLeft: bottom,
-        bottomRight: bottom,
-      );
-
-  static AdBorderRadius horizontal({
-    double left,
-    double right,
-  }) =>
-      AdBorderRadius(
-        topLeft: left,
-        bottomLeft: left,
-        topRight: right,
-        bottomRight: right,
-      );
+Widget buildAndroidPlatformView(Map<String, dynamic> params, String viewType,
+    [bool useHybridComposition = false]) {
+  assert(useHybridComposition != null);
+  final gestures = <Factory<OneSequenceGestureRecognizer>>[
+    // Factory<OneSequenceGestureRecognizer>(() => EagerGestureRecognizer()),
+    // Factory<OneSequenceGestureRecognizer>(() => TapGestureRecognizer()),
+    // Factory<OneSequenceGestureRecognizer>(() => LongPressGestureRecognizer()),
+  ].toSet();
+  if (!useHybridComposition)
+    // virtual display
+    return AndroidView(
+      viewType: viewType,
+      creationParamsCodec: StandardMessageCodec(),
+      creationParams: params,
+      gestureRecognizers: gestures,
+    );
+  else
+    // hybrid composition
+    return PlatformViewLink(
+      viewType: viewType,
+      surfaceFactory: (context, controller) {
+        return AndroidViewSurface(
+          controller: controller,
+          gestureRecognizers: gestures,
+          hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+        );
+      },
+      onCreatePlatformView: (PlatformViewCreationParams p) {
+        return PlatformViewsService.initSurfaceAndroidView(
+          id: p.id,
+          viewType: viewType,
+          layoutDirection: TextDirection.ltr,
+          creationParams: params,
+          creationParamsCodec: StandardMessageCodec(),
+        )
+          ..addOnPlatformViewCreatedListener(p.onPlatformViewCreated)
+          // ..setSize(Size(width, height))
+          ..create();
+      },
+    );
 }